chore: remove old menus module
